New log system
Now all the output will be sent to STDOUT (no more log to file). This is
needed in order to interface with the WEBUI which can obtain information
from the back-end script (mch_config.bash) through STDOUT.

Indiviual logs from the expect scripts will stay on indivual files.
Those logs won't be collected  by the Bash script anymore.

In addition to that, a new message formating was introduced. There are
many different output modes: (1) raw messages (without any color
formating); (2) user friendly messages (colourful); and (3) messages with an
interface for the WEBUI. The main thought about the later is to pass
dynamic information to the web interface. In that way, the web interface
can show dynamic information to the end user, i.e. status of the
process, errors, ...