Added centralized configuration file for Expect scripts
Expect scripts can read some configuration parameters from
a file from now.
There's is also a new argument to set the timestamp for the
generated logs by Expect. In this way, the wrapper script in
Bas can easily grab the files from a run and send them to
Jira.

ICSHWI-2520