New option to configure a MCH through telnet
If -x options is used, the port number won't be taken  into account by
the script. This option is expected to be used when an operator wants to
configure a MCH that is not connected to a MOXA hub.
This option may break a reliable behave of the scripts, so please, use a
MOXA whenever it's possible.