script: Added helper Python script to register MCHs in CSEntry
A new feature is requested to register all the MCHs in CSEntry
before passing the tests. This is needed because the tool is
hosted now int the technical network subnet (172.x.y.z).
Therefore, while running the tests, a valid IP for the MCHs
is needed in that subnet (no access from subnet 172 to
10.x.y.z).

This Python script needs a CSEntry API token in order to access
the API. This token should be hardcoded by now, so bear in
mind that the token variable must be filled with a valid token
at the deployment stage of the MCH configuration tool.

[ICSHWI-4448]