script: Detect if running under the INFRA deployment
INFRA is deploying Python under a virtualenv. In order to access Python3
and the CSEntry API, that environment must be loaded before. The script
may receive an environmental variable when it is executed in the server.
That way, the script will load the virtual environment. Otherwise,
everything must be system wide available.